stagefright : Make iFrameInterval float
I frame interval can be set a fractional by some app. iFrameInterval's
storage type is changed to float

+++ media/libstagefright/include/media/stagefright/ACodec.h
@@ -570,7 +570,7 @@ protected:
 
     virtual void setBFrames(OMX_VIDEO_PARAM_MPEG4TYPE * /*mpeg4type*/) {}
     virtual void setBFrames(OMX_VIDEO_PARAM_AVCTYPE * /*h264type*/,
-        const int32_t /*iFramesInterval*/, const int32_t /*frameRate*/) {}
+        const float /*iFramesInterval*/, const int32_t /*frameRate*/) {}
     virtual status_t getVQZIPInfo(const sp<AMessage> & /*msg*/) {
         return OK;
     }
